codance/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.22
python-dotenv==1.0.0
numpy==1.26.0
pandas==2.1.1
python-multipart==0.0.6
opencv-python==4.8.1.78
scikit-learn==1.3.1
tensorflow==2.14.0
pytorch==2.1.0
scipy==1.11.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pytest==7.4.2
httpx==0.25.0 